+++ app/src/main/kotlin/org/fossify/gallery/activities/EditActivity.kt
@@ -8,7 +8,7 @@ import android.net.Uri
 import android.os.Bundle
 import android.os.Handler
 import android.provider.MediaStore
-import android.widget.RelativeLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.graphics.scale
 import androidx.core.net.toUri
 import androidx.exifinterface.media.ExifInterface
@@ -26,6 +26,7 @@ import com.zomato.photofilters.FilterPack
 import com.zomato.photofilters.imageprocessors.Filter
 import org.fossify.commons.dialogs.ColorPickerDialog
 import org.fossify.commons.extensions.applyColorFilter
+import org.fossify.commons.extensions.setFillWithStroke
 import org.fossify.commons.extensions.beGone
 import org.fossify.commons.extensions.beGoneIf
 import org.fossify.commons.extensions.beVisible
@@ -130,8 +131,15 @@ class EditActivity : BaseCropActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
+        isCropIntent = intent.extras?.get(CROP) == "true"
         setupEdgeToEdge(
-            padBottomSystem = listOf(binding.bottomEditorPrimaryActions.root)
+            padBottomSystem = listOf(
+                if (isCropIntent) {
+                    binding.bottomEditorCropRotateActions.root
+                } else {
+                    binding.bottomEditorPrimaryActions.root
+                }
+            )
         )
 
         if (checkAppSideloading()) {
@@ -210,10 +218,15 @@ class EditActivity : BaseCropActivity() {
             else -> uri!!
         }
 
-        isCropIntent = extras?.get(CROP) == "true"
         if (isCropIntent) {
             binding.bottomEditorPrimaryActions.root.beGone()
-            (binding.bottomEditorCropRotateActions.root.layoutParams as RelativeLayout.LayoutParams).addRule(RelativeLayout.ALIGN_PARENT_BOTTOM, 1)
+
+            val params = binding.bottomEditorCropRotateActions.root.layoutParams as? ConstraintLayout.LayoutParams
+            if (params != null) {
+                params.bottomToBottom = binding.activityEditHolder.id
+                binding.bottomEditorCropRotateActions.root.layoutParams = params
+            }
+
             binding.editorToolbar.menu.findItem(R.id.overwrite_original).isVisible = false
         }
 
@@ -273,7 +286,7 @@ class EditActivity : BaseCropActivity() {
                     dataSource: DataSource,
                     isFirstResource: Boolean
                 ): Boolean {
-                    ColorModeHelper.setColorModeForImage(this@EditActivity, bitmap)
+                    ColorModeHelper.setColorModeForImage(this@EditActivity, bitmap, config.ultraHdrRendering)
                     val currentFilter = getFiltersAdapter()?.getCurrentFilter()
                     if (filterInitialBitmap == null) {
                         loadCropImageView()
@@ -801,7 +814,8 @@ class EditActivity : BaseCropActivity() {
 
     private fun updateDrawColor(color: Int) {
         drawColor = color
-        binding.bottomEditorDrawActions.bottomDrawColor.applyColorFilter(color)
+        binding.bottomEditorDrawActions.bottomDrawColor
+            .setFillWithStroke(color, getProperBackgroundColor())
         config.lastEditorDrawColor = color
         binding.editorDrawCanvas.updateColor(color)
     }